[step2] refactor: Controller 패키지 이동 및 클래스 네이밍 변경
